+++ src/core/functions/generate-style.ts
@@ -1,22 +1,28 @@
 import validStyleProps from "../../utils/valid-style-props";
 import BaseComponentProps from "../interfaces/base-components-props";
+import MediaStyleProps from "../interfaces/media-style-props";
 import StyleProps from "../interfaces/style-props";
 
 const generateStyle = (stylingProps: BaseComponentProps) => {
   //variables that will store the split styling ex: "fs-14px w-200px" => ["fs-14px", "w-200px"]
   let splitStyle: string[] = [];
   let splitHoverStyle: string[] = [];
+  let splitFocusStyle: string[] = [];
 
   //Object that stores the final style detected from the styling strings on the component
   const finalStyle: StyleProps = {};
 
-  //if the dev specified the base style and hover style strings we split them
+  //if the dev specified the base style,hover and focus style strings we split them
   if (stylingProps.style) {
     splitStyle = stylingProps.style.split(";");
   }
 
   if (stylingProps.hover) {
     splitHoverStyle = stylingProps.hover.split(";");
+  }
+
+  if (stylingProps.focus) {
+    splitFocusStyle = stylingProps.focus.split(";");
   }
 
   //We loop through all the array of styles provided ex : ["fs-14px", "w-200px"]
@@ -56,6 +62,51 @@ const generateStyle = (stylingProps: BaseComponentProps) => {
           splitStyleComponent[1];
     }
   });
+
+  //We initialize focus in order to not get error on trying to access a value from a non
+  //existing property
+  finalStyle["focus"] = {};
+
+  //We do the same thing we did on base style, but with focus
+  splitFocusStyle.forEach((styleComponent: string) => {
+    const splitStyleComponent: string[] = styleComponent.split("-");
+
+    if (splitStyleComponent[0].trim() in validStyleProps) {
+      const styleName = validStyleProps[splitStyleComponent[0].trim()];
+
+      if (finalStyle["focus"])
+        finalStyle["focus"][styleName as keyof StyleProps] =
+          splitStyleComponent[1];
+    }
+  });
+
+  //if the dev specified media queries styling
+  if (stylingProps.media) {
+    finalStyle["media"] = {};
+    //we loop through all the media style strings
+    for (let mediaBreakPoint in stylingProps.media) {
+      if (stylingProps.media[mediaBreakPoint as keyof MediaStyleProps]) {
+        finalStyle["media"][mediaBreakPoint as keyof MediaStyleProps] = {};
+        stylingProps.media[mediaBreakPoint as keyof MediaStyleProps]
+          ?.split(";")
+          .forEach((styleComponent: string) => {
+            const splitStyleComponent: string[] = styleComponent.split("-");
+
+            if (splitStyleComponent[0].trim() in validStyleProps) {
+              const styleName = validStyleProps[splitStyleComponent[0].trim()];
+
+              if (finalStyle.media) {
+                const mediaBreakPointKey =
+                  finalStyle["media"][mediaBreakPoint as keyof MediaStyleProps];
+                if (mediaBreakPointKey)
+                  mediaBreakPointKey[styleName as keyof StyleProps] =
+                    splitStyleComponent[1];
+              }
+            }
+          });
+      }
+    }
+  }
 
   //if bgImage prop is provided
   if (stylingProps.bgImage) finalStyle.backgroundImage = stylingProps.bgImage;