+++ src/core/interfaces/base-components-props.ts
@@ -1,8 +1,10 @@
+import MediaStyleProps from "./media-style-props";
+
 interface BaseComponentProps {
   style?: string;
   hover?: string;
   focus?: string;
-  media?: string;
+  media?: MediaStyleProps;
   bgImage?: string;
   background?: string;
   bgPos?: string;