+++ src/core/interfaces/media-style-props.ts
@@ -1,0 +1,9 @@
+interface MediaStyleProps {
+  sm?: string;
+  md?: string;
+  lg?: string;
+  xl?: string;
+  "2xl"?: string;
+}
+
+export default MediaStyleProps;