+++ src/index.ts
@@ -1,10 +1,12 @@
 import Button from "./components/button";
+import Box from "./components/box";
 import SkypieaUICore from "./core/skypiea-ui-core";
 import SkypieaUIReducer from "./core/skypiea-ui-reducer";
 import SkypieaUIContext from "./core/skypiea-ui-context";
 import useDesignSystem from "./hooks/useDesignSystem";
 
 export { Button };
+export { Box };
 export { SkypieaUICore };
 export { SkypieaUIReducer };
 export { SkypieaUIContext };