+++ src/utils/default-breakpoints.ts
@@ -1,0 +1,9 @@
+const DefaultBreakPoints = {
+  sm: "640px",
+  md: "768px",
+  lg: "1024px",
+  xl: "1280px",
+  "2xl": "1536px",
+};
+
+export default DefaultBreakPoints;